use_google_api_key/pyproject.toml
[project]
name = "use_google_api_key"
version = "0.1.0"
description = "Simple project to use Gemini API with google-genai SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "jupyter",
]
